Begin moving files for network package migration
